+++ Microshoppy.Sales/Microshoppy.Sales/src/CQRS/Command/UpdateSalesProductCommand.cs
@@ -8,10 +8,10 @@ namespace Microshoppy.Sales.CQRS.Command
 	{
 		public Guid ProductId { get; set; }
 
-		public double Cost { get; set; }
+		public decimal Cost { get; set; }
 
-		public double NetPrice { get; set; }
+		public decimal NetPrice { get; set; }
 
-		public double TaxPercentage { get; set; }
+		public decimal TaxPercentage { get; set; }
 	}
 }
